osa4

+++ controllers/blogs.js
@@ -1,40 +1,53 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
 
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username:1, name: 1})
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title, 
     author: body.author, 
     url: body.url, 
     likes: body.likes === undefined ? 0 : body.likes,
+    user: user._id
   })
 
-  try {
-    const savedBlog = await blog.save()
-    response.json(savedBlog.toJSON())
-  } catch (error) {
-    response.status(400).send(error.message)
-  }
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
-  //const savedBlog = await blog.save()
-  //response.json(savedBlog.toJSON())
+  response.json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (error) {
-    response.status(404).send(error.message)
-  }
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).json({error: 'deletion unauthorized'})
+    }
+
+    
 })
